Refactoring: changing the structure of an infrastructure project
* the IdentityException file was moved to the Infrastructure\Identity\Exceptions folder

* the EmailService file was moved to the Infrastructure\EmailService folder. Thus,
the file with the service class (EmailService. cs) and the file with the client's smpt
configuration class (EmailOptions. cs)are located in the same folder

* the ITokenClaimService.cs files and the IdentityTokenClaimService.cs
file were moved to the Infrastructure\Identity folder closer to the place of
use. This creates an Identity component that will later be moved to a
separate project
