--- serverapp/Infrastructure/Exceptions/IdentityException.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Infrastructure.Exceptions
-{
-    public class IdentityException:Exception
-    {
-        public IdentityException()
-        {
-
-        }
-
-        public string ErrorMessage { get; set; }
-    }
-}
